Move REST runtime to _rest
